% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apicomplexa.proj}
\alias{apicomplexa.proj}
\title{Apicomplexa geophytter+ topology simplex}
\format{An object of class \code{data.frame} with 252 rows and 6 columns.}
\usage{
apicomplexa.proj
}
\description{
Apicomplexa geophytter+ topology simplex
}
\keyword{datasets}

