% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apicomplexa.tops}
\alias{apicomplexa.tops}
\title{Apicomplexa geophytter+ topology simplex}
\format{An object of class \code{geophytter} of length 2.}
\usage{
apicomplexa.tops
}
\description{
Apicomplexa geophytter+ topology simplex
}
\keyword{datasets}

