% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{as.hexagons}
\alias{as.hexagons}
\title{Expand hexagon centerpoint grid into hexagon vertices}
\usage{
as.hexagons(data, formula = y ~ x, groupvar = "cell")
}
\arguments{
\item{data}{dataframe}

\item{formula}{formula defining the columns with centerpoint data}

\item{groupvar}{name for new column with hexagon group index}
}
\value{
a new data frame with 6*nrow(data) rows and ncol(data)+1 columns
}
\description{
This function expands a data frame of hexagonal tiling centerpoints into one containing hexagon vertices
suitable for plotting with \code{\link{geom_polygon}}. The columns containing the x and y
coordinates of the centerpoints are specified in the formula, \code{y~x}. The groupvar parameter
names a new column in the output containing the hexagon id (i.e., \code{aes(group=groupvar)}).
}
\examples{
lungfish.hex <- as.hexagons(lungfish.tops$simplex)
}
\author{
Grady Weyenberg
}

