% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{coord_simplex}
\alias{coord_simplex}
\title{Special case of coord_fixed with aspect ratio set}
\usage{
coord_simplex(...)
}
\arguments{
\item{...}{passed to \link{coord_fixed}}
}
\description{
Special case of coord_fixed with aspect ratio set
}
\author{
Grady Weyenberg
}

