% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{corner_labels}
\alias{corner_labels}
\title{Add labels to the corners of the simplex}
\usage{
corner_labels(offset = 0.02, ...)
}
\arguments{
\item{offset}{distance to offset the labels from the corners}

\item{...}{passed to \link{geom_text}}
}
\description{
Add labels to the corners of the simplex
}
\author{
Grady Weyenberg
}

