% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count.topologies}
\alias{count.topologies}
\title{Count the trees in x which match the topologies in y}
\usage{
count.topologies(x, y)
}
\arguments{
\item{x}{list of phylo or multiPhylo of trees to count (analogous to first parameter of \code{\link{match}}).}

\item{y}{trees with topologies to be matched against (second parameter of \code{\link{match}}).}
}
\description{
Count the trees in x which match the topologies in y
}
\author{
Grady Weyenberg
}

