% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot.geophytter}
\alias{ggplot.geophytter}
\title{Convenience method for plotting the topology simplex}
\usage{
\method{ggplot}{geophytter}(data, mapping = aes_string(x = "x", y = "y", group
  = "cell", fill = "topology_index"), ..., environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link{fortify}}. If not specified,
must be suppled in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be suppled in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{environment}{If an variable defined in the aesthetic mapping is not
found in the data, ggplot will look for it in this environment. It defaults
to using the environment in which \code{ggplot()} is called.}
}
\description{
Convenience method for plotting the topology simplex
}
\examples{
library(ggplot2)

ggplot(lungfish.tops) +
  corner_labels() +
  geom_point(aes(x=x,y=y), lungfish.proj, inherit.aes = FALSE, size=0.5)

ggplot(lungfish.tops, show.legend=FALSE) +
  scale_fill_grey(start=0.5, end=0.9) +
  corner_labels() +
  geom_point(aes(x=x,y=y), lungfish.proj, inherit.aes = FALSE, size=0.5) +
  geom_label(aes(x=x,y=y,label=topology_index),
    aggregate(cbind(x,y)~topology_index, lungfish.tops$simplex, mean),
    inherit.aes = FALSE)
}
\author{
Grady Weyenberg
}

