% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lungfish.proj}
\alias{lungfish.proj}
\title{Lungfish sample projections onto locus of Frechet means}
\format{a dataframe
\describe{
 \item{squared_projected_dist}{the squared distance to the locus of Frechet means}
 \item{p1,p2,p3}{the weight vector of the projected point}
 \item{x,y}{cartesian coordinates for plotting the weight vector}
}}
\usage{
lungfish.proj
}
\description{
The location of the projections of the lungfish sample trees onto the locus of
Frechet means described by \code{\link{lungfish.tops}}.
}
\keyword{datasets}

