% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lungfish.tops}
\alias{lungfish.tops}
\title{Lungfish geophytter+ topology simplex}
\format{a geophytter object
\describe{
 \item{simplex}{the simplex dataframe}
 \item{trees}{the tree topologies}
}}
\usage{
lungfish.tops
}
\description{
The topology simplex for the second-order treespace PCA produced by geophytter+
}
\keyword{datasets}

