% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.projections}
\alias{read.projections}
\title{Read geophytter+ projections file}
\usage{
read.projections(file)
}
\arguments{
\item{file}{filename}
}
\value{
a dataframe with the squared projected distance, and the simplex coordinates and their projection into cartesian space.
}
\description{
Read geophytter+ projections file
}
\examples{
filename <- system.file("extdata", "lungfish_projections.txt.gz",
                        package="geophytterplus.plot")
lungfish.proj <- read.projections(filename)
}
\author{
Grady Weyenberg
}

