% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.topologies}
\alias{read.topologies}
\title{Read geophytter+ simplex topologies file}
\usage{
read.topologies(file)
}
\arguments{
\item{file}{filename to read}
}
\value{
a list with elements simplex and trees (see Details)
}
\description{
Reads the file and returns a list with the simplex and the tree topologies.
The simplex is a data frame with the probability weight vector, the tree topology index, and the (x,y) projection of the simplex onto cartesian coordinates.
The tree topologies is an ape::multiPhylo object with the trees corresponding to the indices in the simplex dataframe.
}
\examples{
filename <- system.file("extdata", "lungfish_topologies.txt.gz",
                        package="geophyttertools")
lungfish.tops <- read.topologies(filename)
}
\author{
Grady Weyenberg
}

