% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{treegrid}
\alias{treegrid}
\title{Plot tree topologies from geophytter object}
\usage{
treegrid(x, ..., mar = c(1.1, 1.1, 1.1, 1.1))
}
\arguments{
\item{x}{geophytter object}

\item{...}{additional parameters passed to ape::plot.phylo}

\item{mar}{margins for the plots}
}
\description{
A convenience wrapper for \code{\link{plot.phylo}} which plots the
tree topologies contained in the geophytter+ locus of Frechet mean
simplex. Plots the trees contained in \code{x$trees} in a grid, using
\code{main=names($tree)} as the main titles for the tree plots.
}
\examples{
treegrid(lungfish.tops, type="c", direction="d")
}
\author{
Grady Weyenberg
}

